box: Split up `Box`, `Flex` and `Stack` (#1168)
