+++ packages/react/flex/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "dist/ag.ds-next-react-flex.cjs.js"
+}
