+++ packages/react/src/box/docs/code.mdx
@@ -1,7 +1,3 @@
 ## Props
 
 <ComponentPropsTable name="Box" />
-
-<ComponentPropsTable name="Flex" />
-
-<ComponentPropsTable name="Stack" />
