+++ packages/react/src/flex/docs/code.mdx
@@ -1,0 +1,3 @@
+## Props
+
+<ComponentPropsTable name="Flex" />
