+++ packages/react/src/stack/docs/code.mdx
@@ -1,0 +1,3 @@
+## Props
+
+<ComponentPropsTable name="Stack" />
