+++ packages/react/stack/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "dist/ag.ds-next-react-stack.cjs.js"
+}
